api/requirements.txt
Flask==2.2.5
Flask-Cors==4.0.0
boto3==1.33.13
python-dotenv==0.21.1
marshmallow==3.19.0
